refactor(math): move test-only fft helpers (get_powers_of_primitive_root, compose_fft) into the test tree
